+++ sdk/@launchdarkly/observability-android/lib/src/main/kotlin/com/launchdarkly/observability/replay/plugin/SessionReplay.kt
@@ -1,11 +1,8 @@
 package com.launchdarkly.observability.replay.plugin
 
 import com.launchdarkly.observability.BuildConfig
-import com.launchdarkly.observability.interfaces.LDExtendedInstrumentation
-import com.launchdarkly.observability.plugin.InstrumentationContributor
-import com.launchdarkly.observability.plugin.InstrumentationContributorManager
-import com.launchdarkly.observability.replay.ReplayInstrumentation
 import com.launchdarkly.observability.replay.ReplayOptions
+import com.launchdarkly.observability.replay.SessionReplayService
 import com.launchdarkly.observability.sdk.LDObserve
 import com.launchdarkly.observability.sdk.LDReplay
 import com.launchdarkly.sdk.android.LDClient
@@ -13,6 +10,7 @@ import com.launchdarkly.sdk.android.integrations.EnvironmentMetadata
 import com.launchdarkly.sdk.android.integrations.Hook
 import com.launchdarkly.sdk.android.integrations.Plugin
 import com.launchdarkly.sdk.android.integrations.PluginMetadata
+import com.launchdarkly.sdk.android.integrations.RegistrationCompleteResult
 import timber.log.Timber
 import java.util.Collections
 
@@ -23,12 +21,12 @@ import java.util.Collections
  */
 class SessionReplay(
     private val options: ReplayOptions = ReplayOptions(),
-) : Plugin(), InstrumentationContributor {
+) : Plugin() {
 
-    private var cachedInstrumentations: List<LDExtendedInstrumentation>? = null
+    private val sessionReplayHook = SessionReplayHook()
 
     @Volatile
-    var replayInstrumentation: ReplayInstrumentation? = null
+    var sessionReplayService: SessionReplayService? = null
 
     override fun getMetadata(): PluginMetadata {
         return object : PluginMetadata() {
@@ -38,28 +36,29 @@ class SessionReplay(
     }
 
     override fun register(client: LDClient, metadata: EnvironmentMetadata?) {
-        System.out.println("LD:SessionReplay:register LDObserve.context= ${LDObserve.context}")
-        LDObserve.context?.let {
-            InstrumentationContributorManager.add(client, this)
-        } ?: run {
+        val context = LDObserve.context ?: run {
             Timber.tag(TAG).e("Observability plugin is not initialized")
+            return
         }
-    }
 
-    override fun provideInstrumentations(): List<LDExtendedInstrumentation> = synchronized(this) {
-        val instrumentations = cachedInstrumentations ?: LDObserve.context?.let { context ->
-            val instrumentation = ReplayInstrumentation(options, context).also { replayInstrumentation = it }
-            listOf(instrumentation).also { cachedInstrumentations = it }
-        }.orEmpty()
+        if (LDReplay.client != null) {
+            Timber.tag(TAG).e("Session Replay instance already exists")
+            return
+        }
 
-        replayInstrumentation?.let(LDReplay::init)
-        instrumentations
+        val service = SessionReplayService(options, context)
+        LDReplay.init(service)
+        sessionReplayService = service
+        sessionReplayHook.delegate = service
+    
     }
 
     override fun getHooks(metadata: EnvironmentMetadata?): MutableList<Hook> {
-        return Collections.singletonList(
-            SessionReplayHook(this)
-        )
+        return Collections.singletonList(sessionReplayHook)
+    }
+
+    override fun onPluginsReady(result: RegistrationCompleteResult?, metadata: EnvironmentMetadata?) {
+        sessionReplayService?.initialize()
     }
 
     companion object {